Improved error reporting

+++ src/main/java/no/home/automation/ws/action/UpdateDeviceAction.java
@@ -3,6 +3,7 @@ package no.home.automation.ws.action;
 import java.util.HashMap;
 import java.util.Map;
 
+import no.home.automation.dao.DeviceRowMapper;
 import no.home.automation.model.DefaultHandler;
 import no.home.automation.model.Device;
 import no.home.automation.model.UpdateDeviceRequest;
@@ -18,6 +19,7 @@ import com.google.gson.GsonBuilder;
 public class UpdateDeviceAction extends DefaultHandler<UpdateDeviceRequest, UpdateDeviceResponse>
 {
 	private JdbcTemplate	jdbcTemplate	= null;
+	private String			message			= "OK";
 
 	public UpdateDeviceAction(boolean mustBeAuthenticated, JdbcTemplate jdbcTemplate)
 	{
@@ -36,7 +38,7 @@ public class UpdateDeviceAction extends DefaultHandler<UpdateDeviceRequest, Upda
 		else if (request.getType() == TYPE.RENAME)
 			result = updateDevice(request.getDevice());
 
-		return new UpdateDeviceResponse(request.getDevice(), result);
+		return new UpdateDeviceResponse(request.getDevice(), message, result);
 	}
 
 	@Override
@@ -51,11 +53,19 @@ public class UpdateDeviceAction extends DefaultHandler<UpdateDeviceRequest, Upda
 	{
 		try
 		{
+			Device existingDevice = findDevice(device.getSensorId(), device.getUnitCode());
+			if (existingDevice == null)
+			{
+				message = "There is no such device";
+				return false;
+			}
+
 			jdbcTemplate.update("DELETE FROM device WHERE id=?", device.getId());
 			return true;
 		}
 		catch (Exception ex)
 		{
+			message = "Could not delete device";
 			logger.error("Could not delete device", ex);
 			return false;
 		}
@@ -65,11 +75,19 @@ public class UpdateDeviceAction extends DefaultHandler<UpdateDeviceRequest, Upda
 	{
 		try
 		{
+			Device existingDevice = findDevice(device.getSensorId(), device.getUnitCode());
+			if (existingDevice == null)
+			{
+				message = "There is no such device";
+				return false;
+			}
+
 			jdbcTemplate.update("UPDATE device SET name=? WHERE id=?", device.getName(), device.getId());
 			return true;
 		}
 		catch (Exception ex)
 		{
+			message = "Could not update device";
 			logger.error("Could not update device", ex);
 			return false;
 		}
@@ -79,6 +97,13 @@ public class UpdateDeviceAction extends DefaultHandler<UpdateDeviceRequest, Upda
 	{
 		try
 		{
+			Device existingDevice = findDevice(device.getSensorId(), device.getUnitCode());
+			if (existingDevice != null)
+			{
+				message = "Device already exists, rename instead";
+				return false;
+			}
+
 			SimpleJdbcInsert simpleInsert = new SimpleJdbcInsert(jdbcTemplate);
 			simpleInsert.withTableName("device");
 			simpleInsert.setGeneratedKeyName("id");
@@ -95,8 +120,22 @@ public class UpdateDeviceAction extends DefaultHandler<UpdateDeviceRequest, Upda
 		}
 		catch (Exception ex)
 		{
+			message = "Could not insert device";
 			logger.error("Could not insert device", ex);
 			return false;
+		}
+	}
+
+	Device findDevice(int sensorId, int unitCode)
+	{
+		try
+		{
+			return jdbcTemplate.queryForObject("SELECT * FROM device WHERE SensorId=? and UnitCode=?", new DeviceRowMapper(), sensorId, unitCode);
+		}
+		catch (Exception ex)
+		{
+			logger.error("Could not find device", ex);
+			return null;
 		}
 	}
 }
